requirements.txt
asgiref==3.2.10
backcall==0.2.0
beautifulsoup4==4.6.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
Django==3.0.7
django-appconf==1.0.4
django-bootstrap4==2.1.1
django-extensions==3.0.1
django-imagekit==4.0.2
django-sslserver==0.22
future==0.18.2
google-api-core==1.21.0
google-auth==1.18.0
google-cloud-core==1.3.0
google-cloud-speech==1.3.2
google-cloud-storage==1.29.0
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
grpcio==1.30.0
idna==2.10
imageio==2.8.0
imageio-ffmpeg==0.4.2
importlib-metadata==1.6.1
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.1
JPype1==0.7.5
kiwisolver==1.2.0
konlpy==0.5.2
lxml==4.5.1
matplotlib==3.2.2
moviepy==1.0.3
numpy==1.19.0
oauthlib==3.1.0
opencv-python==4.2.0.34
parso==0.7.0
pickleshare==0.7.5
pilkit==2.0
Pillow==7.1.2
proglog==0.1.9
prompt-toolkit==3.0.5
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
tqdm==4.47.0
traitlets==4.3.3
tweepy==3.8.0
urllib3==1.25.9
wcwidth==0.2.5
wordcloud==1.7.0
zipp==3.1.0
